Fix code formatting
Near the end of the file, the closing triple-backtick delimiter on one of the code blocks was not shunted to the next line, which was causing markdown parsing to fail and render the remaining portion of the document as a code block.

This pull request basically just moves the closing delimiter of the code block to the next line to make sure that everything renders correctly. It also places the command flush with the first column, but that's the less important change for sure.